+++ Presentation.Web/Controllers/API/AuthorizeController.cs
@@ -14,6 +14,7 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Web;
 using System.Web.Helpers;
+using Core.ApplicationServices;
 using Core.ApplicationServices.Authentication;
 using Core.ApplicationServices.Organizations;
 using Core.DomainModel.Result;
