+++ Presentation.Web/Controllers/API/PasswordResetRequestController.cs
@@ -2,6 +2,7 @@
 using System.Net;
 using System.Net.Http;
 using System.Web.Http;
+using Core.ApplicationServices;
 using Core.DomainModel;
 using Core.DomainServices;
 using Infrastructure.Services.Cryptography;
