+++ Presentation.Web/Controllers/OData/UsersController.cs
@@ -5,6 +5,7 @@ using Core.DomainServices;
 using Presentation.Web.Infrastructure.Attributes;
 using System.Linq;
 using System.Web.Http;
+using Core.ApplicationServices;
 using Core.ApplicationServices.Authorization.Permissions;
 using Infrastructure.Services.Types;
 using Microsoft.AspNet.OData;
